refactor: encapsulate converter elements and logic into it's own component
